+++ R/bridgesampling.R
@@ -81,7 +81,6 @@ bridge_sampler.brmsfit <- function(samples, ...) {
       "usable in method 'bridge_sampler'."
     )
   }
-  require_backend("rstan", samples)
   # otherwise bridge_sampler might not work in a new R session
   samples <- update_misc_env(samples)
   out <- try(bridge_sampler(samples$fit, ...))